requirements-dev.txt
-r requirements.txt
ruff>=0.9.10,<1.0
pre-commit>=4.0,<5.0
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
mypy>=1.14,<2.0
django-stubs>=6.0.1,<7.0
types-Pillow>=10.2,<13.0
